# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Axon Server API",
    "description": "API for Axon Server http services.",
    "version": "2025.0.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8024",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v2/endpoints": {
      "get": {
        "tags": [
          "Integration Endpoints"
        ],
        "summary": "List all endpoints",
        "description": "Retrieves all integration endpoints with their handlers and connection status",
        "operationId": "endpoints",
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "List of endpoints retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EndpointOverview"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Integration Endpoints"
        ],
        "summary": "Register a new endpoint",
        "description": "Creates a new integration endpoint with the provided configuration",
        "operationId": "registerEndpoint",
        "parameters": [
          {
            "name": "context",
            "in": "query",
            "description": "Context name",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Context name",
              "example": "default"
            },
            "example": "default"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Endpoint"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Endpoint successfully created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Endpoint": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "commandUrl": {
            "type": "string"
          },
          "queryUrl": {
            "type": "string"
          },
          "eventUrl": {
            "type": "string"
          },
          "healthUrl": {
            "type": "string"
          },
          "wrappingType": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "metaData": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "CommandHandlerOverview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "loadFactor": {
            "type": "integer",
            "format": "int32"
          },
          "commandUrl": {
            "type": "string"
          }
        }
      },
      "EndpointOverview": {
        "type": "object",
        "properties": {
          "context": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "commandUrl": {
            "type": "string"
          },
          "queryUrl": {
            "type": "string"
          },
          "eventUrl": {
            "type": "string"
          },
          "healthUrl": {
            "type": "string"
          },
          "wrappingType": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "metaData": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "commandHandlers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandHandlerOverview"
            }
          },
          "eventHandlers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventHandlerOverview"
            }
          },
          "queryHandlers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryHandlerOverview"
            }
          },
          "connections": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EventHandlerOverview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "batchSize": {
            "type": "integer",
            "format": "int32"
          },
          "filter": {
            "type": "string"
          },
          "sequencingPolicy": {
            "type": "string"
          },
          "sequencingPolicyParameters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eventUrl": {
            "type": "string"
          }
        }
      },
      "QueryHandlerOverview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "queryUrl": {
            "type": "string"
          }
        }
      },
...
    }
  }
}